Refactor xml feed job posting importer
